+++ src/main/java/org/springframework/core/io/Resource.java
@@ -1,23 +1,16 @@
 package org.springframework.core.io;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URL;
 
 /**
- * 资源的抽象接口
+ * 资源的抽象和访问接口
  *
  * @author derekyi
  * @date 2020/11/25
  */
 public interface Resource {
 
-	boolean exists();
-
-	File getFile() throws IOException;
-
 	InputStream getInputStream() throws IOException;
 
-	URL getURL() throws IOException;
 }